backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["necklace17 <niklas.scholz@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
django-rest-swagger = "^2.2.0"
PyYAML = "^6.0"
gunicorn = "^20.1.0"
django-filter = "^21.1"
django-cors-headers = "^3.13.0"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pylint = "^2.14.1"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
